requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.13.0
cycler==0.12.1
decorator==5.2.1
deepmultilingualpunctuation==1.0.1
defusedxml==0.7.1
einops==0.8.1
executing==2.2.1
ffmpeg-python==0.2.0
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.0
h11==0.14.0
h5py==3.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.35.0
idna==3.10
# ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
# networkx==3.5
# numpy==2.3.3
opencv-python==4.11.0.86
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==4.21.2
psutil==7.1.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-settings>=1.0.0
Pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.38.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.4
wcwidth==0.2.13
webvtt-py==0.5.1
Werkzeug==3.1.3
youtube-search-python==1.6.6
youtube-transcript-api==0.6.0
yt-dlp==2025.9.5
